backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
python-dotenv
xgboost
pandas
scikit-learn
tensorflow
Pillow
